+++ src/components/rankings/OfficialRanking.js
@@ -1,9 +1,8 @@
 import React, {Component} from "react";
 import {Helmet} from "react-helmet";
-import {connect} from "react-redux";
 import {Table, Flag, Responsive, Icon} from "semantic-ui-react";
-import * as dateUtil from "../util/date.js";
-import countries from "./countries";
+import * as dateUtil from "../../util/date.js";
+import countries from "../common/countries";
 
 class OfficialRanking extends Component {
