+++ src/components/rankings/PointRanking.js
@@ -1,7 +1,7 @@
 import React, {Component} from "react";
 import {Helmet} from "react-helmet";
 import { Table, Flag, Responsive, Icon } from 'semantic-ui-react'
-import countries from "./countries";
+import countries from "../common/countries";
 
 class PointRanking extends Component {
     constructor(props) {
