+++ src/components/rankings/Rankings.js
@@ -1,14 +1,14 @@
 import React, {Component} from "react";
 import {Header} from "semantic-ui-react";
 import {Helmet} from "react-helmet";
-import * as riderStageConnectionsActions from "../actions/riderStageConnectionsActions";
-import store from "../store";
+import * as riderStageConnectionsActions from "../../actions/riderStageConnectionsActions";
+import store from "../../store";
 import {Route, Switch, Link, Redirect} from 'react-router-dom';
 import {Menu} from 'semantic-ui-react';
-import OfficialRankingContainer from '../containers/OfficialRankingContainer'
-import VirtualRankingContainer from '../containers/VirtualRankingContainer'
-import PointRankingContainer from '../containers/PointRankingContainer'
-import MountainRankingContainer from '../containers/MountainRankingContainer'
+import OfficialRankingContainer from '../../containers/OfficialRankingContainer'
+import VirtualRankingContainer from '../../containers/VirtualRankingContainer'
+import PointRankingContainer from '../../containers/PointRankingContainer'
+import MountainRankingContainer from '../../containers/MountainRankingContainer'
 
 class Rankings extends Component {
