+++ src/components/rankings/VirtualRanking.js
@@ -1,8 +1,8 @@
 import React, {Component} from "react";
 import {Helmet} from "react-helmet";
 import { Table, Flag, Responsive, Icon } from 'semantic-ui-react';
-import * as dateUtil from "../util/date.js";
-import countries from "./countries";
+import * as dateUtil from "../../util/date.js";
+import countries from "../common/countries";
 
 class VirtualRanking extends Component {
     constructor(props) {
