+++ src/components/search/SearchResult.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react'
-import RiderInfo from "./RiderDetail/RiderInfo";
-import RiderJudgment from "./RiderDetail/RiderJudgment";
-import RiderTrikot from "./RiderDetail/RiderTrikots";
-import RiderRaceGroup from "./RiderDetail/RiderRaceGroup";
+import RiderInfo from "../RiderDetail/RiderInfo";
+import RiderJudgment from "../RiderDetail/RiderJudgment";
+import RiderTrikot from "../RiderDetail/RiderTrikots";
+import RiderRaceGroup from "../RiderDetail/RiderRaceGroup";
 import {Divider, Icon, Header, Button} from "semantic-ui-react";
-import store from "../store";
-import * as searchActions from "../actions/searchActions";
+import store from "../../store";
+import * as searchActions from "../../actions/searchActions";
 
 class SearchResult extends Component {
     constructor(props){
