+++ src/components/settings/Settings.js
@@ -1,8 +1,8 @@
 import React, {Component} from "react";
 import {Header, Button, Divider, Form} from "semantic-ui-react";
 import {Helmet} from "react-helmet";
-import * as settingsActions from '../actions/settingsActions'
-import store from '../store';
+import * as settingsActions from '../../actions/settingsActions'
+import store from '../../store';
 
 class Settings extends Component {
     constructor(props) {
