+++ src/components/tricots/TricotsStart.js
@@ -1,13 +1,11 @@
 import React, {Component} from "react";
 import {Helmet} from "react-helmet";
-import * as maillotActions from "../actions/maillotActions";
-import store from "../store";
-import SingleTrikotActualContainer from "../containers/SingleTrikotActualContainer";
+import * as maillotActions from "../../actions/maillotActions";
+import store from "../../store";
+import SingleTrikotContainer from "../../containers/SingleTrikotContainer";
 import {List} from "semantic-ui-react";
-import * as riderStageConnectionsActions from "../actions/riderStageConnectionsActions";
 
-
-class TricotsActual extends Component {
+class TricotsStart extends Component {
 
     constructor(props){
         super(props);
@@ -19,7 +17,6 @@ class TricotsActual extends Component {
 
     fetchCurrentMaillots(id) {
         store.dispatch(maillotActions.getCurrentMaillots(id));
-        store.dispatch(riderStageConnectionsActions.getRiderStageConnectionsFromAPI(id));
         this.setState({updated: true});
     }
 
@@ -36,14 +33,14 @@ class TricotsActual extends Component {
         return(
             <div className="App-Content">
                 <Helmet>
-                    <title>Trikotträger Aktuell</title>
+                    <title>Trikotträger Start</title>
                 </Helmet>
                 <List divided verticalAlign='middle' size="massive">
-                    {maillots.map(x => <SingleTrikotActualContainer key={x.id} data={x}/>)}
+                    {maillots.map(x => <SingleTrikotContainer key={x.id} data={x}/>)}
                 </List>
             </div>
         );
     }
 }
 
-export default TricotsActual;
+export default TricotsStart;
