+++ src/components/views/TrackView.js
@@ -1,10 +1,9 @@
 import React, {Component} from "react";
 import {Helmet} from "react-helmet";
 import {Timeline, TimelineEvent} from 'react-event-timeline';
-import {connect} from "react-redux";
-import RiderSearch from "./RiderSearch";
-import SearchResult from "./SearchResult";
-import TimeLineEndBlock from "./TimeLineEndBlock";
+import SearchResult from "../search/SearchResult";
+import TimeLineEndBlock from "../common/TimeLineEndBlock";
+import RiderSearchContainer from '../../containers/RiderSearchContainer'
 
 class TrackView extends Component {
     render() {
@@ -37,7 +36,7 @@ class TrackView extends Component {
                             })}
                         </Timeline>
                         <TimeLineEndBlock content="START"/>
-                        <RiderSearch/>
+                        <RiderSearchContainer/>
                     </div>
                 }
                 {search.displayResult === true &&
