+++ src/containers/OfficialRankingContainer.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import OfficialRanking from "../components/OfficialRanking";
+import OfficialRanking from "../components/rankings/OfficialRanking";
 
 function mapStateToProps(store) {
   return {
