+++ src/containers/SettingsContainer.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import Settings from "../components/Settings";
+import Settings from "../components/settings/Settings";
 
 function mapStateToProps(store) {
   return {
