+++ src/containers/TricotsActualContainer.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import TricotsActual from "../components/TricotsActual";
+import TricotsActual from "../components/tricots/TricotsActual";
 
 
 function mapStateToProps(store) {
